+++ src/cdm/SettingsModel.ts
@@ -11,6 +11,7 @@ export type FilterGroupCondition = {
     condition: string;
     disabled: boolean;
     filters: FilterGroup[];
+    label?: string;
 }
 export type AtomicFilter = {
     field: string;
@@ -32,6 +33,9 @@ export interface LocalSettings {
     cell_size: string;
     current_row_template: string;
     group_folder_column: string;
+    remove_empty_folders: boolean;
+    hoist_files_with_empty_attributes: boolean;
+    automatically_group_files: boolean;
     frontmatter_quote_wrap: boolean;
     pagination_size: number;
     remove_field_when_delete_column: boolean;
@@ -45,6 +49,7 @@ export interface LocalSettings {
     source_data: string;
     sticky_first_column: boolean;
     row_templates_folder: string;
+    enable_js_formulas: boolean;
     formula_folder_path: string;
     inline_default: boolean;
     inline_new_position: string;