+++ src/components/HeaderMenu.tsx
@@ -9,9 +9,12 @@ import { ColumnSettingsModal } from "components/modals/columnSettings/ColumnSett
 import { TableColumn } from "cdm/FolderModel";
 import { HeaderActionResponse } from "cdm/HeaderActionModel";
 import { HeaderMenuProps } from "cdm/HeaderModel";
+import { destination_folder } from "helpers/FileManagement";
+import { FileGroupingService } from "services/FileGroupingService";
 
 const HeaderMenu = (headerMenuProps: HeaderMenuProps) => {
   const { table, column } = headerMenuProps.headerProps;
+  const { view } = table.options.meta;
   const [columnsInfo, columnActions] = table.options.meta.tableState.columns(
     (state) => [state.info, state.actions]
   );
@@ -21,6 +24,8 @@ const HeaderMenu = (headerMenuProps: HeaderMenuProps) => {
   const configInfo = table.options.meta.tableState.configState(
     (state) => state.info
   );
+  const configActions =  table.options.meta.tableState.configState((state) => state.actions);
+  const ddbbConfig = table.options.meta.tableState.configState( (store) => store.ddbbConfig);
 
   /** Header props */
   const {
@@ -140,6 +145,17 @@ const HeaderMenu = (headerMenuProps: HeaderMenuProps) => {
           configInfo.getLocalSettings()
         );
       });
+      // Rename column in group_folder_column
+      const groupFolderColumn =
+        ddbbConfig.group_folder_column.split(",");
+      if (groupFolderColumn.includes(column.columnDef.id)) {
+        const newGroupFolderColumn = groupFolderColumn
+          .map((item) => item === column.columnDef.id? newKey : item)
+          .join(",");
+        configActions.alterConfig({ group_folder_column: newGroupFolderColumn });
+        // Reorganize files and remove empty folders
+        dataActions.groupFiles()
+      }
   }
 
   function handleKeyDown(e: any) {