+++ src/components/NavBar.tsx
@@ -11,18 +11,18 @@ import InsertDriveFileIcon from "@mui/icons-material/InsertDriveFile";
 import SettingsIcon from "@mui/icons-material/Settings";
 import Menu from "@mui/material/Menu";
 import MenuItem from "@mui/material/MenuItem";
-
 import Toolbar from "@mui/material/Toolbar";
 import { c } from "helpers/StylesHelper";
 import EditFiltersButton from "components/reducers/DataviewFilters";
 import { MenuButtonStyle } from "components/styles/NavBarStyles";
 import { SettingsModal } from "Settings";
-import CsvReader from "./navbar/CsvReader";
+import CsvReader from "components/navbar/CsvReader";
 import { t } from "lang/helpers";
 import AppBar from "@mui/material/AppBar";
 import ButtonGroup from "@mui/material/ButtonGroup";
-import ToggleFiltersButton from "./reducers/ToggleFiltersButton";
+import ToggleFiltersButton from "components/reducers/ToggleFiltersButton";
 import Paper from "@mui/material/Paper";
+import QuickFilters from "components/reducers/QuickFilters";
 
 export function NavBar(navBarProps: NavBarProps) {
   const { table } = navBarProps;
@@ -143,6 +143,16 @@ export function NavBar(navBarProps: NavBarProps) {
             <ToggleFiltersButton table={table} />
             <EditFiltersButton table={table} />
           </Paper>
+
+          <Box
+            sx={{
+              overflowX: "auto",
+              display: "flex",
+              padding: { xs: "0", md: "5px" },
+            }}
+          >
+            <QuickFilters table={table} />
+          </Box>
           <Box sx={{ flexGrow: 1 }} />
           <Box
             justifyContent={"flex-start"}