+++ src/components/headerActions/handlers/buttons/RemoveColumnHandlerAction.tsx
@@ -6,6 +6,8 @@ import { UpdateRowOptions } from "helpers/Constants";
 import { RowDataType, TableColumn } from "cdm/FolderModel";
 import headerButtonComponent from "components/headerActions/HeaderButtonComponent";
 import { EditEngineService } from "services/EditEngineService";
+import { destination_folder } from "helpers/FileManagement";
+import { FileGroupingService } from "services/FileGroupingService";
 
 export default class RemoveColumnHandlerAction extends AbstractHeaderAction {
   globalHeaderActionResponse: HeaderActionResponse;
@@ -31,6 +33,8 @@ export default class RemoveColumnHandlerAction extends AbstractHeaderAction {
 function removeButton(headerActionResponse: HeaderActionResponse) {
   const { hooks } = headerActionResponse;
   const { column, table } = headerActionResponse.headerMenuProps.headerProps;
+  const { view } = table.options.meta;
+  const configActions =  table.options.meta.tableState.configState((state) => state.actions);
   const ddbbConfig = table.options.meta.tableState.configState(
     (store) => store.ddbbConfig
   );
@@ -63,6 +67,17 @@ function removeButton(headerActionResponse: HeaderActionResponse) {
     dataActions.removeDataOfColumn(column.columnDef as TableColumn);
     columnActions.remove(column.columnDef as TableColumn);
     hooks.setExpanded(false);
+    // Remove column from group_folder_column
+    const groupFolderColumn =
+        ddbbConfig.group_folder_column.split(",");
+    if (groupFolderColumn.includes(column.columnDef.id)) {
+      const newGroupFolderColumn = groupFolderColumn
+        .filter((item) => item !== column.columnDef.id)
+        .join(",");
+      configActions.alterConfig({ group_folder_column: newGroupFolderColumn });
+      // Reorganize files and remove empty folders
+      dataActions.groupFiles();
+    }
   };
 
   return headerButtonComponent({